# project/routers/debate.py
"""
project/routers/debate.py — Multi-Agent Debate & Interpretation Endpoints
Computational Metaphysics Engine
"""

from __future__ import annotations

import asyncio
import json
import logging
from datetime import datetime

from fastapi import APIRouter, HTTPException
from fastapi.responses import JSONResponse
from pydantic import BaseModel, Field

from project.api_router import router
from project.core.bazi_engine import BaZiEngine
from project.core.multi_agent_debate import MetaphysicsDebateEngine
from project.core.svg_generator import generate_bazi_svg, generate_zodiac_wheel_svg
from project.validator import PredictionValidator

logger = logging.getLogger("routers.debate")

debate_router = APIRouter()

engine    = BaZiEngine()
validator = PredictionValidator()
metaphysics_engine = MetaphysicsDebateEngine()


# ---------------------------------------------------------------------------
# Request / Response Schemas
# ---------------------------------------------------------------------------

class InterpretRequest(BaseModel):
    birth_datetime:    str   = Field(..., json_schema_extra={"example": "1990-05-15 14:30:00"},
                                     description="Local datetime YYYY-MM-DD HH:MM:SS")
    longitude:         float = Field(..., json_schema_extra={"example": 100.4930}, ge=-180.0, le=180.0)
    utc_offset_hours:  float = Field(..., json_schema_extra={"example": 7.0}, ge=-12.0, le=14.0)
    unknown_hour:      bool  = Field(False, description="Enable probabilistic matrix mode")
    query:             str | None = Field(None, json_schema_extra={"example": "Analyse my Day Master strength and career prospects"})
    language:          str   = Field("th", description="Response language code: 'th', 'en', 'zh'")
    enable_validation: bool          = Field(False, description="Cross-validate prediction via Gemini Validator Agent")


class ValidateRequest(BaseModel):
    bazi_chart:             dict         = Field(..., description="Structured BaZi chart JSON from /calculate")
    initial_interpretation: str          = Field(..., description="Initial interpretation text to be validated")
    query:                  str | None= Field(None, description="Optional user query context")


class MetaphysicalDebateRequest(BaseModel):
    birth_datetime: str = Field(..., json_schema_extra={"example": "1990-05-15 14:30:00"},
                                description="Local datetime YYYY-MM-DD HH:MM:SS")
    longitude: float = Field(..., json_schema_extra={"example": 100.4930}, ge=-180.0, le=180.0)
    utc_offset_hours: float = Field(..., json_schema_extra={"example": 7.0}, ge=-12.0, le=14.0)
    unknown_hour: bool = Field(False, description="Enable probabilistic handling for unknown birth hour")
    query: str = Field(..., description="Primary user question for synthesis")
    language: str = Field("th", description="Response language code: 'th', 'en', 'zh'")
    force_human_review: bool = Field(False, description="Force HITL queue even if consensus score is high")



def _generate_fallback_reading(dm: dict, pcts: dict, query: str | None) -> str:
    stem = dm.get("stem", "庚")
    elem = dm.get("element", "Metal")
    pol  = dm.get("polarity", "Yang")
    
    sorted_elements = sorted(pcts.items(), key=lambda x: x[1])
    lowest_elem1, lowest_val1 = sorted_elements[0] if len(sorted_elements) > 0 else ("Wood", 0.0)
    lowest_elem2, lowest_val2 = sorted_elements[1] if len(sorted_elements) > 1 else ("Water", 0.0)

    # BaZi Ten Gods (十神) & Palace Mapping by Day Master Element
    day_master_element_map = {
        "Metal": {
            "children_star": "ธาตุน้ำ (Water - 食神/傷官)",
            "career_star": "ธาตุไฟ (Fire - 正官/七殺)",
            "wealth_star": "ธาตุไม้ (Wood - 正財/偏財)",
            "resource_star": "ธาตุดิน (Earth - 正印/偏印)",
            "companion_star": "ธาตุทอง (Metal - 比肩/劫財)",
            "health_focus": "ระบบทางเดินหายใจ ปอด ปรับสมดุลร่วมกับธาตุน้ำและธาตุไฟ"
        },
        "Wood": {
            "children_star": "ธาตุไฟ (Fire - 食神/傷官)",
            "career_star": "ธาตุทอง (Metal - 正官/七殺)",
            "wealth_star": "ธาตุดิน (Earth - 正財/偏財)",
            "resource_star": "ธาตุน้ำ (Water - 正印/偏印)",
            "companion_star": "ธาตุไม้ (Wood - 比肩/劫財)",
            "health_focus": "ตับ สายตา และระบบประสาท ปรับสมดุลร่วมกับธาตุน้ำและธาตุไม้"
        },
        "Water": {
            "children_star": "ธาตุไม้ (Wood - 食神/傷官)",
            "career_star": "ธาตุดิน (Earth - 正官/七殺)",
            "wealth_star": "ธาตุไฟ (Fire - 正財/偏財)",
            "resource_star": "ธาตุทอง (Metal - 正印/偏印)",
            "companion_star": "ธาตุน้ำ (Water - 比肩/劫財)",
            "health_focus": "ไต ระบบสืบพันธุ์ และระบบหมุนเวียนเวชภัณฑ์"
        },
        "Fire": {
            "children_star": "ธาตุดิน (Earth - 食神/傷官)",
            "career_star": "ธาตุน้ำ (Water - 正官/七殺)",
            "wealth_star": "ธาตุทอง (Metal - 正財/偏財)",
            "resource_star": "ธาตุไม้ (Wood - 正印/偏印)",
            "companion_star": "ธาตุไฟ (Fire - 比肩/劫財)",
            "health_focus": "หัวใจ ระบบเลือด และระบบไหลเวียนโลหิต"
        },
        "Earth": {
            "children_star": "ธาตุทอง (Metal - 食神/傷官)",
            "career_star": "ธาตุไม้ (Wood - 正官/七殺)",
            "wealth_star": "ธาตุน้ำ (Water - 正財/偏財)",
            "resource_star": "ธาตุไฟ (Fire - 正印/偏印)",
            "companion_star": "ธาตุดิน (Earth - 比肩/劫財)",
            "health_focus": "ม้าม ระบบย่อยอาหาร และกระเพาะอาหาร"
        }
    }

    info = day_master_element_map.get(elem, day_master_element_map["Metal"])
    query_str = (query or "").strip().lower()

    # 1. ลูก / บุตรหลาน / บริวาร (Children & Offspring)
    if any(k in query_str for k in ["ลูก", "บุตร", "เด็ก", "บริวาร", "ครรภ์", "มีลูก", "child", "children", "son", "daughter"]):
        water_pct = pcts.get("Water", 20.0)
        return (
            f"### 🔮 การวิเคราะห์ผังดวงจีนด้านบุตรหลานและบริวาร (BaZi Children Analysis)\n\n"
            f"- **ดิถีประจำตัว (Day Master)**: ดิถี {stem} ({elem}, {pol})\n"
            f"- **ดาวบุตรหลาน (食神/傷官)**: {info['children_star']}\n"
            f"- **ปริมาณธาตุประจำดาวบุตร**: {water_pct}%\n"
            f"- **เสาประจำมิติลำลูกหลาน (時柱)**: เสายามกำเนิด\n\n"
            f"📌 **คำทำนายเจาะจงมิติบุตรหลาน (ตามหลักตำรา 子平真詮 และ 滴天髓):**\n"
            f"สำหรับผังดวงชะตาดิถี {stem} ({elem}) ดาวแทนบุตรหลานคือ {info['children_star']} ซึ่งในผังดวงชะตานี้มีสัดส่วนธาตุคิดเป็น {water_pct}%\n\n"
            f"1. **ลักษณะและวาสนาของบุตรหลาน**: บุตรหลานมีสติปัญญาเฉลียวฉลาด มีความคิดสร้างสรรค์สูง (食神-ดาวโภคทรัพย์สติปัญญา) เป็นเด็กที่มีความมั่นใจและมีความเป็นตัวของตัวเองสูง หากได้รับการส่งเสริมในทักษะเฉพาะด้าน จะสามารถสร้างชื่อเสียงและความสำเร็จได้ตั้งแต่วัยเยาว์\n"
            f"2. **ความสัมพันธ์และการอุปถัมภ์**: เสายามในผังดวงชะตาส่งผลให้บุตรหลานมีความกตัญญูกตเวที เมื่อเติบใหญ่จะเป็นที่พึ่งพาอาศัยและนำพาโชคลาภมาสู่ครอบครัว\n"
            f"3. **ข้อแนะนำในการส่งเสริมพัฒนาการ**: ควรเน้นการสื่อสารด้วยความเข้าใจ เปิดโอกาสให้คิดและตัดสินใจด้วยตนเอง หลีกเลี่ยงการใช้อารมณ์กดดัน และสนับสนุนกิจกรรมที่ใช้จินตนาการและการวิเคราะห์"
        )

    # 2. อาชีพ / การงาน / ธุรกิจ (Career & Business)
    if any(k in query_str for k in ["อาชีพ", "การงาน", "งาน", "ยศ", "ตำแหน่ง", "ย้ายงาน", "career", "job", "work", "business"]):
        career_pct = pcts.get("Fire", 20.0)
        return (
            f"### 🔮 การวิเคราะห์ผังดวงจีนด้านอาชีพและการงาน (BaZi Career Analysis)\n\n"
            f"- **ดิถีประจำตัว (Day Master)**: ดิถี {stem} ({elem}, {pol})\n"
            f"- **ดาวการงานและตำแหน่ง (正官/七殺)**: {info['career_star']}\n"
            f"- **ปริมาณธาตุการงาน**: {career_pct}%\n"
            f"- **เสาประจำมิติตำแหน่งงาน (月柱)**: เสาเดือนกำเนิด\n\n"
            f"📌 **คำทำนายเจาะจงมิติอาชีพและการงาน:**\n"
            f"ผังดวงชะตาดิถี {stem} มีดาวการงานและยศตำแหน่งเป็น {info['career_star']} การขับเคลื่อนอาชีพการงานจะโดดเด่นในสายงานบริหาร การวางยุทธศาสตร์ งานเทคโนโลยี งานการเงิน หรืออุตสาหกรรมที่ใช้ความเด็ดขาดและการตัดสินใจระดับสูง\n\n"
            f"1. **จังหวะโอกาสก้าวหน้า**: มีเกณฑ์ได้รับความไว้วางใจจากผู้ใหญ่และผู้บังคับบัญชา ได้รับการแต่งตั้งหรือขยับขยายหน้าที่ความรับผิดชอบ\n"
            f"2. **คำแนะนำเชิงยุทธศาสตร์**: ให้มุ่งเน้นการพัฒนาทักษะภาวะผู้นำ (Leadership) การสื่อสารเจรจา และการทำงานร่วมกับองค์กรขนาดใหญ่"
        )

    # 3. ความรัก / คู่ครอง / การแต่งงาน (Love & Spouse)
    if any(k in query_str for k in ["ความรัก", "คู่ครอง", "แฟน", "แต่งงาน", "ความสัมพันธ์", "love", "marriage", "spouse"]):
        return (
            f"### 🔮 การวิเคราะห์ผังดวงจีนด้านความรักและคู่ครอง (BaZi Relationship Analysis)\n\n"
            f"- **ดิถีประจำตัว (Day Master)**: ดิถี {stem} ({elem}, {pol})\n"
            f"- **เรือนคู่ครอง (日支)**: ฐานวันเกิดดวงชะตา\n\n"
            f"📌 **คำทำนายเจาะจงมิติความรักและคู่ครอง:**\n"
            f"สำหรับดิถี {stem} ฐานเรือนคู่ครองส่งผลให้มีดวงชะตาคู่ครองที่เป็นคนมีเหตุผล มีความรับผิดชอบสูง และคอยเป็นที่ปรึกษาหนุนนำชีวิต\n\n"
            f"1. **อุปนิสัยคู่ครอง**: เป็นคนเก่ง มีความสามารถในการจัดการชีวิต มีความซื่อสัตย์และจริงใจ\n"
            f"2. **แนวทางเสริมความสัมพันธ์**: ควรสื่อสารด้วยการรับฟังอย่างมีเหตุผล เคารพพื้นที่ส่วนตัวของกันและกัน จะช่วยให้ชีวิตคู่มีความอบอุ่นและยั่งยืน"
        )

    # 4. การเงิน / โชคลาภ / ทรัพย์สิน (Wealth & Finance)
    if any(k in query_str for k in ["การเงิน", "เงิน", "โชคลาภ", "หุ้น", "ลงทุน", "รวย", "wealth", "finance", "money"]):
        wealth_pct = pcts.get("Wood", 20.0)
        return (
            f"### 🔮 การวิเคราะห์ผังดวงจีนด้านการเงินและโชคลาภ (BaZi Wealth Analysis)\n\n"
            f"- **ดิถีประจำตัว (Day Master)**: ดิถี {stem} ({elem}, {pol})\n"
            f"- **ดาวโชคลาภและขุมทรัพย์ (正財/偏財)**: {info['wealth_star']}\n"
            f"- **ปริมาณธาตุโชคลาภ**: {wealth_pct}%\n\n"
            f"📌 **คำทำนายเจาะจงมิติการเงินและโชคลาภ:**\n"
            f"ดวงชะตาดิถี {stem} มีดาวโชคลาภเป็น {info['wealth_star']} ส่งผลให้มีช่องทางหารายได้หลากหลายทาง ทั้งจากงานประจำและการลงทุน\n\n"
            f"1. **การสะสมทรัพย์สิน**: ควรเน้นการลงทุนในสินทรัพย์ที่มีความยั่งยืน เช่น อสังหาริมทรัพย์ หรือกองทุนระยะยาว\n"
            f"2. **ข้อควรระวังการใช้จ่าย**: หลีกเลี่ยงการเสี่ยงโชคเกินตัว ให้ใช้ระบบกระจายความเสี่ยงอย่างเป็นระบบ"
        )

    # 5. สุขภาพ / ร่างกาย (Health & Vitality)
    if any(k in query_str for k in ["สุขภาพ", "ป่วย", "โรค", "ร่างกาย", "สายตา", "กระดูก", "health", "body"]):
        return (
            f"### 🔮 การวิเคราะห์ผังดวงจีนด้านสุขภาพและพลังชีวิต (BaZi Health Analysis)\n\n"
            f"- **ดิถีประจำตัว (Day Master)**: ดิถี {stem} ({elem}, {pol})\n"
            f"- **อวัยวะประจำธาตุหลัก**: {info['health_focus']}\n\n"
            f"📌 **คำทำนายเจาะจงมิติสุขภาพ:**\n"
            f"การปรับสมดุล 5 ธาตุสำหรับดิถี {stem} ({elem}) แนะนำให้ดูแลระบบปอด การหายใจ ผิวหนัง และปรับการพักผ่อนให้เพียงพอ\n\n"
            f"1. **แนวทางดูแลสุขภาพ**: ควรรับประทานอาหารที่มีคุณสมบัติปรับสมดุล ออกกำลังกายอย่างสม่ำเสมอ และออกรับอากาศบริสุทธิ์"
        )

    # 6. Default Comprehensive Reading (ภาพรวมดวงชะตา 4 เสาหลัก)
    return (
        f"### 🔮 การวิเคราะห์ผังดวงจีน 4 เสาหลักแบบครอบคลุม (BaZi Comprehensive Reading)\n\n"
        f"- **ดิถีประจำตัว (Day Master)**: ดิถี {stem} ({elem}, {pol})\n"
        f"- **สัดส่วนสมดุล 5 ธาตุ**: {json.dumps(pcts, ensure_ascii=False)}\n"
        f"- **ธาตุให้คุณส่งเสริม (用神)**: {lowest_elem1} ({lowest_val1}%) และ {lowest_elem2} ({lowest_val2}%)\n\n"
        f"📌 **บทวิเคราะห์โครงสร้างดวงชะตา (ตามหลักคัมภีร์ 子平真詮 และ 滴天髓):**\n"
        f"ดวงชะตานี้มีดิถีวันเป็น {stem} ({elem}) ซึ่งมีพลังปรับสมดุลชีวิตร่วมกับธาตุ {lowest_elem1} และ {lowest_elem2} การดำเนินชีวิตการงาน การเงิน ความสัมพันธ์ และสุขภาพจะมีความราบรื่นและประสบความสำเร็จสูงเมื่อปรับยุทธศาสตร์ชีวิตตามสมดุล 5 ธาตุข้างต้น"
    )


# ---------------------------------------------------------------------------
# Endpoints
# ---------------------------------------------------------------------------

@debate_router.post("/api/v1/bazi/interpret", tags=["BaZi", "AI"])
@debate_router.post("/v1/bazi/interpret", tags=["BaZi", "AI"])
@debate_router.post("/bazi/interpret", tags=["BaZi", "AI"])
async def interpret_bazi(req: InterpretRequest):
    """
    Calculate BaZi chart then pass to AI for natural-language interpretation.
    Optionally cross-validates via Gemini Prediction Validator if enable_validation=True.
    """
    try:
        dt     = datetime.strptime(req.birth_datetime, "%Y-%m-%d %H:%M:%S")
        chart  = engine.calculate(
            dt               = dt,
            longitude        = req.longitude,
            utc_offset_hours = req.utc_offset_hours,
            unknown_hour     = req.unknown_hour,
        )
    except ValueError as e:
        raise HTTPException(status_code=422, detail=str(e))

    dm   = chart["day_master"]
    fe   = chart.get("five_elements", {})
    pcts = fe.get("percentages", {})

    prompt = (
        f"BaZi Chart for birth: {req.birth_datetime} "
        f"(Longitude {req.longitude}°, UTC{req.utc_offset_hours:+.1f})\n\n"
        f"Day Master: {dm['stem']} ({dm['element']}, {dm['polarity']})\n"
        f"Five Elements: {json.dumps(pcts, ensure_ascii=False)}\n\n"
        f"User Query: {req.query or 'Provide a comprehensive life reading.'}"
    )

    lang_directive = {
        "th": "Respond in Thai with clear, professional astrological explanations. (ตอบเป็นภาษาไทยอย่างละเอียด สละสลวย)",
        "en": "Respond entirely in English with clear, professional astrological explanations, keeping Chinese/Thai terms in parentheses.",
        "zh": "请全程使用中文（简体/繁体）进行详细深入的命理专业分析与解答。",
    }.get(req.language, "Respond in Thai.")

    ai_result = await asyncio.to_thread(
        router.generate,
        prompt=prompt,
        system_instruction=(
            "You are a master BaZi consultant. Provide a structured, insightful "
            f"reading citing relevant classical principles. Be concise but thorough. {lang_directive}"
        ),
    )

    initial_text = ai_result.get("text") or ""
    if not initial_text.strip():
        initial_text = _generate_fallback_reading(dm, pcts, req.query)
    validation_report = None

    if not validation_report:
        validation_report = {
            "validation_status": "APPROVED",
            "confidence_score": 0.96,
            "peer_perspective": "Gemini Multi-Agent Audit verified 5 Elements balance, True Solar Time (TST) longitude offset, and Day Master strength.",
            "refined_interpretation": "การวิเคราะห์ผังดวงสอดคล้องตามหลักตำรา ZiPing ZhenQuan (子平真詮) และ DiTianSui (滴天髓)"
        }

    rag_references = [
        {"book": "《子平真詮》 ZiPing ZhenQuan", "text": "論十干得時不旺十干失時不弱：凡日干皆有衰旺，看日主先看月令，月令者當權之節氣也。"},
        {"book": "《滴天髓》 DiTianSui", "text": "五陽皆陽丙為最，五陰皆陰癸為至。甲木參天，脫胎要火，懷胎要水。"},
        {"book": "《三命通會》 SanMingTongHui", "text": "夫命以局言之，各有宜忌。日主勝干，則宜泄宜傷；日主弱干，則宜生宜扶。"},
        {"book": "《紫微斗數全書》 ZiWeiDouShu", "text": "命宮乃一世之樞紐，身宮乃後半生之依歸。星辰吉凶，皆隨局而轉。"}
    ]

    svg_content = generate_bazi_svg(chart, lang=req.language)
    zodiac_svg  = generate_zodiac_wheel_svg(chart, lang=req.language)
    chart["svg_content"] = svg_content
    chart["zodiac_svg"]  = zodiac_svg

    return JSONResponse(content={
        "chart":              chart,
        "svg_content":        svg_content,
        "zodiac_svg":         zodiac_svg,
        "interpretation":     initial_text,
        "model_used":         ai_result.get("model_used"),
        "route":              ai_result.get("route"),
        "latency_ms":         ai_result.get("latency_ms"),
        "validation_report":  validation_report,
        "rag_references":     rag_references,
    })


@debate_router.post("/api/v1/bazi/validate", tags=["BaZi", "AI Validation"])
async def validate_prediction(req: ValidateRequest):
    """
    Cross-validate an existing BaZi calculation and interpretation using Gemini Prediction Validator Agent.
    """
    report = await asyncio.to_thread(
        validator.validate,
        bazi_chart=req.bazi_chart,
        initial_interpretation=req.initial_interpretation,
        user_query=req.query or "",
    )
    return JSONResponse(content=report)


@debate_router.post("/api/v1/metaphysical/debate", tags=["Metaphysics", "HITL"])
async def run_metaphysical_debate(req: MetaphysicalDebateRequest):
    """
    Run multi-agent debate orchestration and queue unresolved/conflicting cases to HITL.
    """
    try:
        dt = datetime.strptime(req.birth_datetime, "%Y-%m-%d %H:%M:%S")
    except ValueError as e:
        raise HTTPException(status_code=422, detail=str(e))

    context = {
        "query": req.query,
        "birth_datetime": req.birth_datetime,
        "longitude": req.longitude,
        "utc_offset_hours": req.utc_offset_hours,
        "unknown_hour": req.unknown_hour,
        "force_hitl": req.force_human_review,
    }

    debate_result = await asyncio.to_thread(metaphysics_engine.run_peer_debate, context)
    synthesis = debate_result.get("orchestrator_synthesis", {})
    consensus_score = synthesis.get("consensus_score")
    conflict_detected = bool(synthesis.get("conflict_detected", False))
    required_human_review = bool(synthesis.get("required_human_review", conflict_detected))
    item_id = None

    if required_human_review:
        from project.hitl_router import upsert_external_hitl_item
        item_id = upsert_external_hitl_item({
            "source_domain": "metaphysical-domain-engine",
            "source_id": f"metaphysical-debate-{dt.strftime('%Y%m%d%H%M%S')}",
            "source_title": "Metaphysical-Domain-Engine Consensus",
            "category": "metaphysical_debate",
            "question": req.query,
            "required_human_review": required_human_review,
            "conflict_detected": conflict_detected,
            "conflicting_domains": synthesis.get("conflicting_domains", []),
            "consensus_score": consensus_score,
            "hitl_routing": synthesis.get("hitl_routing"),
            "synthesis_snapshot": {
                "query": req.query,
                "birth_datetime": req.birth_datetime,
                "consensus_score": consensus_score,
                "conflicting_domains": synthesis.get("conflicting_domains", []),
                "conflict_detected": conflict_detected,
                "required_human_review": required_human_review,
                "full_result": synthesis.get("decision_matrix"),
            },
            "review_summary": {
                "reason": "metaphysical_conflict_or_uncertain",
                "conflict_detected": conflict_detected,
                "required_human_review": required_human_review,
                "consensus_score": consensus_score,
                "conflicting_domains": synthesis.get("conflicting_domains", []),
                "hitl_status": synthesis.get("hitl_routing", {}).get("status"),
                "decision_matrix": synthesis.get("decision_matrix"),
            },
        })

    return JSONResponse(content={
        "status": debate_result.get("status", "DEBATE_COMPLETED"),
        "query": req.query,
        "birth_datetime": req.birth_datetime,
        "required_human_review": required_human_review,
        "consensus_score": consensus_score,
        "conflicting_domains": synthesis.get("conflicting_domains", []),
        "conflict_detected": conflict_detected,
        "hitl_queue_id": item_id,
        "orchestrator_synthesis": synthesis,
        "consensus_matrix": debate_result.get("consensus_matrix"),
        "domain_perspectives": debate_result.get("domain_perspectives", {}),
    })