# project/tests/goldens/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Computational Metaphysics Engine",
    "description": "Modular 10-Domain Metaphysical Calculation, AI Debate & Multi-Agent Engine",
    "version": "1.0.0.34833bc"
  },
  "paths": {
...
    "/bazi/interpret": {
      "post": {
        "tags": [
          "BaZi",
          "AI"
        ],
        "summary": "Interpret Bazi",
        "description": "Calculate BaZi chart then pass to AI for natural-language interpretation.\nOptionally cross-validates via Gemini Prediction Validator if enable_validation=True.",
        "operationId": "interpret_bazi_bazi_interpret_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterpretRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/bazi/interpret": {
      "post": {
        "tags": [
          "BaZi",
          "AI"
        ],
        "summary": "Interpret Bazi",
        "description": "Calculate BaZi chart then pass to AI for natural-language interpretation.\nOptionally cross-validates via Gemini Prediction Validator if enable_validation=True.",
        "operationId": "interpret_bazi_v1_bazi_interpret_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterpretRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bazi/interpret": {
      "post": {
        "tags": [
          "BaZi",
          "AI"
        ],
        "summary": "Interpret Bazi",
        "description": "Calculate BaZi chart then pass to AI for natural-language interpretation.\nOptionally cross-validates via Gemini Prediction Validator if enable_validation=True.",
        "operationId": "interpret_bazi_api_v1_bazi_interpret_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterpretRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bazi/validate": {
      "post": {
        "tags": [
          "BaZi",
          "AI Validation"
        ],
        "summary": "Validate Prediction",
        "description": "Cross-validate an existing BaZi calculation and interpretation using Gemini Prediction Validator Agent.",
        "operationId": "validate_prediction_api_v1_bazi_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/metaphysical/debate": {
      "post": {
        "tags": [
          "Metaphysics",
          "HITL"
        ],
        "summary": "Run Metaphysical Debate",
        "description": "Run multi-agent debate orchestration and queue unresolved/conflicting cases to HITL.",
        "operationId": "run_metaphysical_debate_api_v1_metaphysical_debate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MetaphysicalDebateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InterpretRequest": {
        "properties": {
          "birth_datetime": {
            "type": "string",
            "title": "Birth Datetime",
            "description": "Local datetime YYYY-MM-DD HH:MM:SS",
            "example": "1990-05-15 14:30:00"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "example": 100.493
          },
          "utc_offset_hours": {
            "type": "number",
            "maximum": 14.0,
            "minimum": -12.0,
            "title": "Utc Offset Hours",
            "example": 7.0
          },
          "unknown_hour": {
            "type": "boolean",
            "title": "Unknown Hour",
            "description": "Enable probabilistic matrix mode",
            "default": false
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query",
            "example": "Analyse my Day Master strength and career prospects"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "description": "Response language code: 'th', 'en', 'zh'",
            "default": "th"
          },
          "enable_validation": {
            "type": "boolean",
            "title": "Enable Validation",
            "description": "Cross-validate prediction via Gemini Validator Agent",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "birth_datetime",
          "longitude",
          "utc_offset_hours"
        ],
        "title": "InterpretRequest"
      },
...
      "MetaphysicalDebateRequest": {
        "properties": {
          "birth_datetime": {
            "type": "string",
            "title": "Birth Datetime",
            "description": "Local datetime YYYY-MM-DD HH:MM:SS",
            "example": "1990-05-15 14:30:00"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "example": 100.493
          },
          "utc_offset_hours": {
            "type": "number",
            "maximum": 14.0,
            "minimum": -12.0,
            "title": "Utc Offset Hours",
            "example": 7.0
          },
          "unknown_hour": {
            "type": "boolean",
            "title": "Unknown Hour",
            "description": "Enable probabilistic handling for unknown birth hour",
            "default": false
          },
          "query": {
            "type": "string",
            "title": "Query",
            "description": "Primary user question for synthesis"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "description": "Response language code: 'th', 'en', 'zh'",
            "default": "th"
          },
          "force_human_review": {
            "type": "boolean",
            "title": "Force Human Review",
            "description": "Force HITL queue even if consensus score is high",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "birth_datetime",
          "longitude",
          "utc_offset_hours",
          "query"
        ],
        "title": "MetaphysicalDebateRequest"
      },
...
      "ValidateRequest": {
        "properties": {
          "bazi_chart": {
            "additionalProperties": true,
            "type": "object",
            "title": "Bazi Chart",
            "description": "Structured BaZi chart JSON from /calculate"
          },
          "initial_interpretation": {
            "type": "string",
            "title": "Initial Interpretation",
            "description": "Initial interpretation text to be validated"
          },
          "query": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Query",
            "description": "Optional user query context"
          }
        },
        "type": "object",
        "required": [
          "bazi_chart",
          "initial_interpretation"
        ],
        "title": "ValidateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}